install.bat, changed router.bat

+++ distribution/disttemplates/router/scripts/router.bat
@@ -1,16 +1,22 @@
 @echo off
-where java >nul 2>nul
-if %errorlevel%==1 (
-    @echo Please install Java 11 or later before using SwiftMQ! From Java 15 onwards you need GraalVM: https://graalvm.org
-    exit
-)
+if exists .executables (
+    set /p EXECUTABLES=<.executables
+    set /p JAVA_HOME=<.javahome
 
-for /f tokens^=2-5^ delims^=.-_^" %%j in ('java -fullversion 2^>^&1') do set "jver=%%j"
-if %jver% LSS 11 (
-    @echo Please install Java 11 or later before using SwiftMQ! From Java 15 onwards you need GraalVM: https://graalvm.org
-    exit
+) else (
+    where java >nul 2>nul
+    if %errorlevel%==1 (
+        @echo Please install Java 11 or later before using SwiftMQ! From Java 15 onwards you need GraalVM: https://graalvm.org
+        exit
+    )
+
+    for /f tokens^=2-5^ delims^=.-_^" %%j in ('java -fullversion 2^>^&1') do set "jver=%%j"
+    if %jver% LSS 11 (
+        @echo Please install Java 11 or later before using SwiftMQ! From Java 15 onwards you need GraalVM: https://graalvm.org
+        exit
+    )
+    for /f "tokens=*" %%i in ('java -cp ../jars/swiftmq.jar com.swiftmq.JavaHome') do set JAVAHOME=%%i
 )
-for /f "tokens=*" %%i in ('java -cp ../jars/swiftmq.jar com.swiftmq.JavaHome') do set JAVAHOME=%%i
 set OPENS=--add-opens=java.desktop/java.awt.font=ALL-UNNAMED --add-opens=java.base/java.text=ALL-UNNAMED --add-opens=java.base/java.lang.reflect=ALL-UNNAMED --add-opens=java.base/java.util=ALL-UNNAMED --add-opens=java.base/java.nio=ALL-UNNAMED  --add-opens=java.base/sun.nio.ch=ALL-UNNAMED  --add-opens=java.base/sun.net.dns=ALL-UNNAMED -Dnashorn.args=--no-deprecation-warning
 
 set PRECONFIG=-Dswiftmq.preconfig=../data/preconfig/upgrade-to-12.1.0.xml
@@ -29,4 +35,7 @@ IF DEFINED SWIFTMQ_JVMPARAM (
   set JVMPARAM=%SWIFTMQ_JVMPARAM%
 )
 
-java -server %JVMPARAM% %ROUTEROPT% -cp ../jars/swiftmq.jar com.swiftmq.Router ../data/config/routerconfig.xml
+if "%EXECUTABLES%" == ""
+    java -server %JVMPARAM% %ROUTEROPT% -cp ../jars/swiftmq.jar com.swiftmq.Router ../data/config/routerconfig.xml
+else
+    %EXECUTABLES%/java -server %JVMPARAM% %ROUTEROPT% -cp ../jars/swiftmq.jar com.swiftmq.Router ../data/config/routerconfig.xml
